+++ CHANGELOG.md
@@ -19,3 +19,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - TypeScript and Hardhat Toolbox setup.
 - `.env` support for private keys and RPC configuration.
 - Comprehensive `.gitignore` for build, cache, and environment files.
+- `DroneIdentityNFT` contract for creation of NFT of the drones
+- Delete `MockDroneIdentityNFT` and update `RoutePermission` contract and tests and env.orig file